service/assessment-service/requirements.txt
fastapi>=0.100.0,<0.105.0
uvicorn[standard]>=0.20.0,<0.25.0
httpx>=0.24.0,<0.26.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.5,<0.1.0
python-dotenv>=1.0.0,<2.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
redis>=4.5.0,<5.0.0
celery>=5.3.0,<6.0.0
